[expo][assets] Register the Expo asset plugin with Metro
Expo comes with an asset plugin that adds the MD5 checksums of asset files to each asset object. This is used when uploading hosted assets.

We need to extend Metro and the RN CLI parameters to support passing in plugins from Expo CLI but for SDK 31 we'll patch RN directly.